[network] Move ProtocolId to wire::handshake
Closes: #3358
Approved by: phlip9
